Move datagen into test_modin
